:fire: Remove MvvmCross on Droid startup
