+++ Toggl.Droid/Startup/TogglApplication.cs
@@ -1,16 +1,23 @@
 using System;
 using Android.App;
 using Android.Runtime;
-using MvvmCross.Droid.Support.V7.AppCompat;
-using Toggl.Core.UI;
-using Toggl.Core.UI.ViewModels;
+using Toggl.Core;
+using Toggl.Networking;
 
 namespace Toggl.Droid
 {
     [Application(AllowBackup = false)]
-    public class TogglApplication : MvxAppCompatApplication<Setup, App<LoginViewModel>>
+    public class TogglApplication : Application
     {
-        public TogglApplication(IntPtr javaReference, JniHandleOwnership transfer) : base(javaReference, transfer)
+        private const ApiEnvironment environment =
+#if USE_PRODUCTION_API
+                        ApiEnvironment.Production;
+#else
+                        ApiEnvironment.Staging;
+#endif
+
+        public TogglApplication(IntPtr javaReference, JniHandleOwnership transfer)
+            : base(javaReference, transfer)
         {
         }
 
@@ -18,6 +25,10 @@ namespace Toggl.Droid
         {
             base.OnCreate();
             Firebase.FirebaseApp.InitializeApp(this);
+
+            var applicationContext = Context;
+            var packageInfo = applicationContext.PackageManager.GetPackageInfo(applicationContext.PackageName, 0);
+            AndroidDependencyContainer.EnsureInitialized(environment, Platform.Giskard, packageInfo.VersionName);
 #if USE_ANALYTICS
             Microsoft.AppCenter.AppCenter.Start(
                 "{TOGGL_APP_CENTER_ID_DROID}",
